server/requirements.txt
asgiref==3.8.1
Django==5.1.6
pillow==12.0.0
python-dotenv==1.0.1
qrcode==8.0
reportlab==4.0.7
sqlparse==0.5.3
typing_extensions==4.12.2
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
PyJWT==2.10.1
django-cors-headers==4.3.1
gunicorn==21.2.0
whitenoise==6.6.0
